requirements.txt
# Core dependencies
python-dotenv==1.0.0
tinydb==4.8.0
requests==2.31.0

# CLI interface
rich==13.7.0
click==8.1.7

# File monitoring

# Torrent client integration
qbittorrent-api==2024.1.58

# Audio metadata
mutagen==1.47.0

# EPUB conversion
# Requires Calibre CLI (system package)
# Ubuntu/Debian: sudo apt install calibre
# Fedora: sudo dnf install calibre
# macOS: brew install calibre
# Windows: Download from https://calibre-ebook.com/download
# Provides: ebook-convert command for PDF to EPUB conversion
